Avoid unknown comm values on process exit
When processing fanotify events after a process exited, its
/proc/pid/comm is not available any more (ESRCH or ENOENT).

If the previous event was for the same pid and it got a comm name,
re-use it instead of "unknown". This will of course not work for lots of
interleaved events from different processes (which would require a
full-blown map implementation).