[#48] Add option -d,--dir
When --dir arguments are given and the directories are few enough, use
dir_watch_mode, where directories are added to the notification group,
rather than mounts or file systems. As directories are created and
moved, they are dynamically added to and removed from the notification
group as needed.

When the number of watched directories is large, warn and fall back to
watching mounts or file systems, then filter events after they are
received.

Fix parse_args for --filter to distinguish between FAN_MOVED_FROM and
FAN_MOVED_TO, since dir_watch_mode can produce events that are marked
with one but not the other.

Filtering on path happens rather late in the process, which increases
cache misses for procname. Compensate by caching two values.